moving spatial_join to the spatial_index module, adding options struct for join
