+++ geopolars/src/spatial_index.rs
@@ -2,10 +2,157 @@ use geo::{
     prelude::BoundingRect, Geometry, Line, LineString, MultiLineString, MultiPoint, MultiPolygon,
     Point, Polygon,
 };
-use polars::prelude::{PolarsError, Series};
+use polars::prelude::{DataFrame, JoinType, NamedFrom, PolarsError, Result, Series};
 use rstar::{RTree, RTreeObject, AABB};
 
-use crate::util::iter_geom;
+use crate::util::{geom_at_index, iter_geom, Predicate};
+
+pub struct SpatialJoinArgs<'a> {
+    pub join_type: JoinType,
+    pub predicate: Predicate,
+    pub l_suffix: &'a str,
+    pub r_suffix: &'a str,
+    pub l_index: Option<&'a SpatialIndex>,
+    pub r_index: Option<&'a SpatialIndex>,
+}
+
+impl<'a> Default for SpatialJoinArgs<'a> {
+    fn default() -> Self {
+        Self {
+            join_type: JoinType::Inner,
+            predicate: Predicate::Intersects,
+            l_suffix: "_left",
+            r_suffix: "_right",
+            l_index: None,
+            r_index: None,
+        }
+    }
+}
+
+pub fn spatial_join(
+    lhs: &DataFrame,
+    rhs: &DataFrame,
+    options: SpatialJoinArgs,
+) -> Result<DataFrame> {
+    use geo::algorithm::{contains::Contains, intersects::Intersects};
+
+    let lhs_geometry = lhs.column("geometry")?;
+    let rhs_geometry = rhs.column("geometry")?;
+
+    let spatial_index_left: SpatialIndex = lhs_geometry.try_into().map_err(|_| {
+        PolarsError::ComputeError(std::borrow::Cow::Borrowed(
+            "Failed to generate the spatial index for the left dataframe",
+        ))
+    })?;
+
+    let spatial_index_right: SpatialIndex = rhs_geometry.try_into().map_err(|_| {
+        PolarsError::ComputeError(std::borrow::Cow::Borrowed(
+            "Failed to generate the spatial index for the right dataframe",
+        ))
+    })?;
+
+    let potential_overlaps = spatial_index_left
+        .r_tree
+        .intersection_candidates_with_other_tree(&spatial_index_right.r_tree);
+
+    let mut left_series: Vec<usize> = vec![];
+    let mut right_series: Vec<usize> = vec![];
+
+    for intersection in potential_overlaps {
+        let (lhs_node, rhs_node) = intersection;
+
+        let lhs_geom = geom_at_index(lhs_geometry, lhs_node.index)?;
+        let rhs_geom = geom_at_index(rhs_geometry, rhs_node.index)?;
+
+        let actual_hit = match (&lhs_geom, &rhs_geom, &options.predicate) {
+            // Points and Polygons
+            (Geometry::Point(point), Geometry::Polygon(poly), _) => poly.contains(point),
+            (Geometry::Polygon(poly), Geometry::Point(point), _) => poly.contains(point),
+
+            // Points and MultiPolygons
+            (Geometry::Point(point), Geometry::MultiPolygon(poly), _) => poly.contains(point),
+            (Geometry::MultiPolygon(poly), Geometry::Point(point), _) => poly.contains(point),
+
+            // Polygon and Polygon
+            (Geometry::Polygon(poly_lhs), Geometry::Polygon(poly_rhs), Predicate::Contains) => {
+                poly_lhs.contains(poly_rhs)
+            }
+            (Geometry::Polygon(poly_lhs), Geometry::Polygon(poly_rhs), Predicate::Intersects) => {
+                poly_lhs.intersects(poly_rhs)
+            }
+
+            // Multi Polygon and Polygon
+            (
+                Geometry::MultiPolygon(poly_lhs),
+                Geometry::Polygon(poly_rhs),
+                Predicate::Contains,
+            ) => poly_lhs.contains(poly_rhs),
+            (
+                Geometry::MultiPolygon(poly_lhs),
+                Geometry::Polygon(poly_rhs),
+                Predicate::Intersects,
+            ) => poly_lhs.intersects(poly_rhs),
+
+            // Polygon and MultiPolygon
+            (
+                Geometry::Polygon(poly_lhs),
+                Geometry::MultiPolygon(poly_rhs),
+                Predicate::Intersects,
+            ) => poly_lhs.intersects(poly_rhs),
+            // Line and Point
+            (Geometry::Line(line), Geometry::Point(point), _) => line.contains(point),
+            (Geometry::Point(point), Geometry::Line(line), _) => line.contains(point),
+
+            // LineString and Point
+            (Geometry::LineString(line), Geometry::Point(point), _) => line.contains(point),
+            (Geometry::Point(point), Geometry::LineString(line), _) => line.contains(point),
+
+            // MultiLineString and Point
+            (Geometry::MultiLineString(line), Geometry::Point(point), _) => line.contains(point),
+            (Geometry::Point(point), Geometry::MultiLineString(line), _) => line.contains(point),
+            _ => false,
+        };
+
+        if actual_hit {
+            left_series.push(lhs_node.index);
+            right_series.push(rhs_node.index);
+        }
+    }
+
+    let lhs_index: Vec<u64> = (0..lhs.shape().0).map(|i| i as u64).collect();
+    let rhs_index: Vec<u64> = (0..rhs.shape().0).map(|i| i as u64).collect();
+
+    let lhs_index = Series::new("lhs_index", lhs_index);
+    let rhs_index = Series::new("rhs_index", rhs_index);
+
+    let lhs_join_series: Vec<u64> = left_series.iter().map(|i| *i as u64).collect();
+    let rhs_join_series: Vec<u64> = right_series.iter().map(|i| *i as u64).collect();
+
+    let lhs_join_series = Series::new("lhs_join", lhs_join_series);
+    let rhs_join_series = Series::new("rhs_join", rhs_join_series);
+
+    let join_df: DataFrame = DataFrame::new(vec![lhs_join_series, rhs_join_series])?;
+
+    let lhs_with_index = lhs.hstack(&[lhs_index])?;
+    let rhs_with_index = rhs.hstack(&[rhs_index])?;
+
+    match options.join_type {
+        JoinType::Inner => {
+            let join_one = lhs_with_index.inner_join(&join_df, ["lhs_index"], ["lhs_join"])?;
+            let join_two = join_one.inner_join(&rhs_with_index, ["rhs_join"], ["rhs_index"])?;
+            let result = join_two.drop("lhs_index")?.drop("rhs_join")?;
+            Ok(result)
+        }
+        JoinType::Left => {
+            let join_one = lhs_with_index.left_join(&join_df, ["lhs_index"], ["lhs_join"])?;
+            let join_two = join_one.left_join(&rhs_with_index, ["rhs_join"], ["rhs_index"])?;
+            let result = join_two.drop("lhs_index")?.drop("rhs_join")?;
+            Ok(result)
+        }
+        JoinType::Outer => unimplemented!(),
+        JoinType::Cross => unimplemented!(),
+    }
+}
 
 #[derive(Debug)]
 pub enum NodeEnvelope {
@@ -98,7 +245,7 @@ impl RTreeObject for TreeNode {
 
 impl TryFrom<Geometry<f64>> for NodeEnvelope {
     type Error = PolarsError;
-    fn try_from(geom: Geometry<f64>) -> Result<Self, Self::Error> {
+    fn try_from(geom: Geometry<f64>) -> std::result::Result<Self, Self::Error> {
         match geom {
             Geometry::Polygon(poly) => Ok(poly.into()),
             Geometry::MultiPolygon(multi_poly) => Ok(multi_poly.into()),
@@ -124,7 +271,7 @@ impl SpatialIndex {}
 impl<'a> TryFrom<&'a Series> for SpatialIndex {
     type Error = PolarsError;
 
-    fn try_from(series: &'a Series) -> Result<Self, Self::Error> {
+    fn try_from(series: &'a Series) -> std::result::Result<Self, Self::Error> {
         let mut r_tree: RTree<TreeNode> = RTree::new();
         for (index, geom) in iter_geom(series).enumerate() {
             let node = TreeNode {
@@ -140,7 +287,7 @@ impl<'a> TryFrom<&'a Series> for SpatialIndex {
 impl TryFrom<Series> for SpatialIndex {
     type Error = PolarsError;
 
-    fn try_from(series: Series) -> Result<Self, Self::Error> {
+    fn try_from(series: Series) -> std::result::Result<Self, Self::Error> {
         let mut r_tree: RTree<TreeNode> = RTree::new();
         for (index, geom) in iter_geom(&series).enumerate() {
             let node = TreeNode {
@@ -155,11 +302,13 @@ impl TryFrom<Series> for SpatialIndex {
 
 #[cfg(test)]
 mod tests {
+    use crate::{
+        geoseries::GeoSeries,
+        spatial_index::{spatial_join, SpatialIndex, SpatialJoinArgs},
+    };
     use geo::{polygon, Geometry, Point, Polygon};
-    use polars::prelude::{PolarsError, Series};
+    use polars::prelude::{DataFrame, JoinType, NamedFrom, PolarsError, Series};
     use rstar::AABB;
-
-    use crate::{geoseries::GeoSeries, spatial_index::SpatialIndex};
 
     #[test]
     fn spatial_index_points() {
@@ -230,5 +379,59 @@ mod tests {
         let indexes: Vec<usize> = in_envelope.map(|node| node.index).collect();
         assert!(indexes.contains(&0));
         assert_eq!(indexes.len(), 1);
+    }
+
+    #[test]
+    fn spatial_join_test() {
+        let points: Vec<Point<f64>> = vec![
+            Point::new(0.0, 10.0),
+            Point::new(1.0, 1.0),
+            Point::new(10.0, 1.0),
+            Point::new(1.0, -1.0),
+            Point::new(0.0, -10.0),
+            Point::new(-1.0, -1.0),
+            Point::new(-10.0, 0.0),
+            Point::new(-1.0, 1.0),
+            Point::new(0.0, 10.0),
+        ];
+
+        let geoms: Vec<Geometry<f64>> = points.into_iter().map(|p| p.into()).collect();
+
+        let point_series = Series::from_geom_vec(&geoms).unwrap();
+        let point_value_series = Series::new("point_values", [1., 2., 3., 4., 5., 6., 7., 8., 9.]);
+        let point_df: DataFrame = DataFrame::new(vec![point_series, point_value_series]).unwrap();
+
+        let polygons: Vec<Geometry<f64>> = vec![Geometry::Polygon(polygon![
+            (x:0., y:0.),
+            (x:20., y:0.),
+            (x:20., y:20.),
+            (x:0., y: 20.)
+        ])];
+
+        let polygon_series = Series::from_geom_vec(&polygons).unwrap();
+        let polygon_label_series = Series::new("string_col", ["test"]);
+
+        let polygon_df: DataFrame =
+            DataFrame::new(vec![polygon_series, polygon_label_series]).unwrap();
+
+        let inner_options = SpatialJoinArgs {
+            join_type: JoinType::Inner,
+            ..Default::default()
+        };
+
+        let inner_result: DataFrame = spatial_join(&point_df, &polygon_df, inner_options).unwrap();
+
+        let left_options = SpatialJoinArgs {
+            join_type: JoinType::Left,
+            ..Default::default()
+        };
+
+        let left_result: DataFrame = spatial_join(&point_df, &polygon_df, left_options).unwrap();
+
+        assert_eq!(inner_result.shape(), (2, 4));
+        assert_eq!(left_result.shape(), (9, 4));
+
+        println!("inner {}", inner_result);
+        println!("left {}", left_result);
     }
 }
